+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Fieldsets now have a `smallerHint` option to match other form field components
-- Global header component now has `defaultContent` and `defaultContentBaseURL` options to allow more consistent global content to be used
+- Global header and footer components now have `defaultContent` and `defaultContentBaseURL` options to allow more consistent global content to be used
 
 ### Changed